Fix maxCursor for Trending
For Trending, in the case $cursor is empty, $this->cursor is updated with __getTtwid so it's different from $cursor (unlike in other Items where $this->cursor is the same as $cursor). If we use $cursor for fromReq, the maxCursor will always be 0. We should use $this->cursor instead so that the client can extract maxCursor as ttwid and later request Trending with that ttwid.